requirements.txt
streamlit
pandas
plotly
numpy
transformers
deep-translator
torch
emoji
wordcloud
matplotlib
google-generativeai